added more plotting | add citatin.cff

+++ nightlightsprocessing/clean_cropped_images.py
@@ -7,12 +7,16 @@ def main():
     print("directories", directories)
 
     for directory in directories:
-        path = f"{filepath}/{directory}"
+        if not directory.startswith("."):
+            path = f"{filepath}/{directory}"
+            sub_directories = os.listdir(path)
 
-        for root, dirs, files in os.walk(path):
-            for name in files:
-                if name.endswith((".aux.xml")):
-                    os.remove(os.path.join(root, name))
+            for location_sub_directory in sub_directories:
+                if not location_sub_directory.startswith("."):
+                    for root, dirs, files in os.walk(f"{path}/{location_sub_directory}"):
+                        for name in files:
+                            if name.endswith((".aux.xml")):
+                                os.remove(os.path.join(root, name))
 
 
 if __name__ == "__main__":
